+++ src/components/chat/chatRoom/ChatInput.tsx
@@ -42,6 +42,7 @@ const ChatInput = ({ onSendMessage, onHeightChange }: Props) => {
     return () => {
       window.removeEventListener('resize', handleResize);
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [message]);
 
   const handleSend = () => {
