+++ src/hooks/useModalHistory.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { usePathname, useSearchParams } from 'next/navigation';
 import { useEffect, useRef } from 'react';
