+++ src/hooks/useTravelForm.ts
@@ -197,6 +197,7 @@ const useTravelForm = () => {
     };
 
     loadInitialData();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return {
