+++ src/hooks/useTravelSchedules.ts
@@ -91,6 +91,7 @@ const useTravelSchedules = (
 
   useEffect(() => {
     initDefaultSchedule();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return {
