+++ src/mocks/librarymocks/svg.js
@@ -1,2 +1,3 @@
+/* eslint-disable */
 export default 'svg';
 export const ReactComponent = 'div';
